Avoid noexcept() problem for purecall test.
When trying to throw from the callback (which is a use case to continue
execution), when inside a noexcept function this fails for
understandable reasons and triggers another terminate() call. To be able
to test CR_CPP_PURECALL separately from this, create a test case that
does not trigger the purecall handler inside a destructor which is by
default noexcept.